Fixes 'no module named emailer' issue running pytest.
If you don't have emailer in your installed packages
and you run `pytest` in the emailer root folder.
Then Pytest will fail with the error 'no module named emailer'

https://stackoverflow.com/a/43559758/92584

I think this is because the tests are not in a module, so they don't look
to the root folder to see if there is a module named emailer.  But
by making the tests a module, then they know to look in the pwd for a
module named emailer. 🤷